Added files via upload
There is an error with simulator for testing email. On stack overflow, people have said that it works on devices and not the simulators. I'd like to try it out at the meeting today if possible. My hackintosh is not detecting my phone so I have not had the chance to try it out on my device yet.